Split up the final application instance views
